+++ apps/ymir-sdl3/src/app/events/emu_event_factory.cpp
@@ -55,6 +55,7 @@ EmuEvent SetDebugTrace(bool enable) {
             ctx.saturn.masterSH2.UseTracer(&ctx.tracers.masterSH2);
             ctx.saturn.slaveSH2.UseTracer(&ctx.tracers.slaveSH2);
             ctx.saturn.SCU.UseTracer(&ctx.tracers.SCU);
+            ctx.saturn.CDBlock.UseTracer(&ctx.tracers.CDBlock);
         }
         devlog::info<grp::base>("Debug tracing {}", (enable ? "enabled" : "disabled"));
     });